+++ src/test/kotlin/no/nav/helse/fritakagp/processing/gravid/krav/GravidKravProcessorTest.kt
@@ -22,6 +22,7 @@ import no.nav.helse.fritakagp.integration.gcp.BucketStorage
 import no.nav.helse.fritakagp.jsonEquals
 import no.nav.helse.fritakagp.processing.BakgrunnsJobbUtils.emptyJob
 import no.nav.helse.fritakagp.processing.BakgrunnsJobbUtils.testJob
+import no.nav.helse.fritakagp.processing.brukernotifikasjon.BrukernotifikasjonJobbdata
 import no.nav.helse.fritakagp.processing.brukernotifikasjon.BrukernotifikasjonProcessor
 import no.nav.helse.fritakagp.processing.gravid.krav.GravidKravProcessor.Companion.brevkode
 import no.nav.helse.fritakagp.processing.gravid.krav.GravidKravProcessor.Companion.dokumentasjonBrevkode
@@ -174,7 +175,7 @@ class GravidKravProcessorTest {
         }
 
         val beskjedJobb = opprettetJobber.find { it.type == BrukernotifikasjonProcessor.JOB_TYPE }
-        assertThat(beskjedJobb?.data).contains(BrukernotifikasjonProcessor.Jobbdata.SkjemaType.GravidKrav.name)
+        assertThat(beskjedJobb?.data).contains(BrukernotifikasjonJobbdata.SkjemaType.GravidKrav.name)
         assertThat(beskjedJobb?.data).contains(krav.id.toString())
     }
